papersman: Add support for doc id's, pass map to an index template.
The idea is to support cross-referencing between the docs. Arbitrary xref
relataions can be supported, e.g. "cites", "extension of", etc., and they
can be handled completely on the level of a template.

"md5" of the doc is supported as an id (and is a primary id), and arbitrary
number of "secondary" id's can be defined using syntax:

ids:
- id1
- id2

The idea behind supporting multiple is that if underlying document changes
to another version/rendition, previous "md5" can be added to "ids" instead,
and no links between documents will be broken. Also, this allows to define
human-readable id's, e.g. author + pubdate or similar.